services: Service provider now manages keywords
Defining actions, cli commands and other things
within the scope of the service definition
is implemented step by step into the system.
While the provider adapt (and become compiler)
the main provider is not able to handle a complete
set of definitions with the new keywords
(as in an PHP- or YAML-File).
With this commit the main Provider now knows
how to delegate each section to the appropriate
provider or compiler.