compat: YAML compat interface to Symfony3 & 4
The YAML component of Symfony4 has a slightly type-safer
interface than the Symfony3 YAML component.
When using the loose (Symfony3 YAML) function signatures
while Symfony4 is installed,
then we get lots of errors to comply with the more rigid
function signature.
We create a proxy-like class that establishes compatibility
to both Symfony versions to fulfill the declaration of each
interfaces.

* No more warnings like
  "Declaration of ... should be compatible with ..."
  (for YAML)
* This differs by PHP-Version which could still lead to Warnings
  when Symfony 3 is installed on PHP >= 7.1