services: Transformed into ProviderNode
The definition/provider currently all take care
of their own structure.
Actually this is a tree and some logic
repeats over all nodes,
so we deal with some duplicate code here.
By introducing the ProviderNodes we can make
all sections easier to extend
and get rid of the duplicate code.